Creating a new Superset auth class
