+++ tests/auth/superset_test.py
@@ -3,9 +3,10 @@ Test username:password authentication mechanism.
 """
 
 from requests_mock.mocker import Mocker
+from pytest_mock import MockerFixture
 from yarl import URL
 
-from preset_cli.auth.password import UsernamePasswordAuth
+from preset_cli.auth.superset import UsernamePasswordAuth, SupersetJWTAuth
 
 
 def test_username_password_auth(requests_mock: Mocker) -> None:
@@ -53,3 +54,30 @@ def test_username_password_auth_no_csrf(requests_mock: Mocker) -> None:
     assert auth.get_headers() == {}
 
     assert requests_mock.last_request.text == "username=admin&password=password123"
+
+
+def test_jwt_auth_superset(mocker: MockerFixture) -> None:
+    """
+    Test the ``JWTAuth`` authentication mechanism for Superset tenant.
+    """
+    auth = SupersetJWTAuth("my-token", URL("https://example.org/"))
+    mocker.patch.object(auth, "get_csrf_token", return_value="myCSRFToken")
+
+    assert auth.get_headers() == {
+        "Authorization": "Bearer my-token",
+        "X-CSRFToken": "myCSRFToken",
+    }
+
+
+def test_get_csrf_token(requests_mock: Mocker) -> None:
+    """
+    Test the get_csrf_token method.
+    """
+    auth = SupersetJWTAuth("my-token", URL("https://example.org/"))
+    requests_mock.get(
+        "https://example.org/api/v1/security/csrf_token/",
+        json={"result": "myCSRFToken"},
+    )
+
+    assert auth.get_csrf_token("my-token") == "myCSRFToken"
+
